move deserializer tests around
