+++ test/deserializer/nestedList.test.ts
@@ -1,5 +1,5 @@
-import { transform } from '../src';
-import nestedMdast from './fixtures/nested-list';
+import { transform } from '../../src';
+import nestedMdast from '../fixtures/nested-list';
 
 it('When the list has paragraph elements', () => {
   expect(nestedMdast.children.map((k) => transform(k))).toMatchSnapshot();
